back-end/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-autoslug==1.9.9
django-jazzmin==3.0.0
djangorestframework==3.15.1
Pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
django-cors-headers==4.3.1
drf-spectacular==0.27.2
whitenoise==6.6.0
gunicorn